+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Changed
 * Pyarrow >=23 required
 * dataset `group` and `filter` optimized
+* example app config: FEDERATED renamed, FILTERS removed
 
 ### Added
 * `first` sorts by rank